Possible to give minutes to expire token

+++ server/src/main/java/com/wcwm/wcwm/util/JwtUtil.java
@@ -18,7 +18,7 @@ public class JwtUtil  {
     private String SECRET_KEY;
 
     @Value("${secret.expire-key-hours}")
-    private Integer EXPIRATION_HOURS;
+    private Float EXPIRATION_HOURS;
 
     public String extractUsername(String token) {
         return extractClaim(token, Claims::getSubject);
@@ -48,7 +48,7 @@ public class JwtUtil  {
 
     private String createToken(Map<String, Object> claims, String subject) {
         return Jwts.builder().setClaims(claims).setSubject(subject).setIssuedAt(new Date(System.currentTimeMillis()))
-                .setExpiration(new Date(System.currentTimeMillis() + EXPIRATION_HOURS  * 100 * 60 * 60 * 10))
+                .setExpiration(new Date(System.currentTimeMillis() + (int)(EXPIRATION_HOURS  * 100 * 60 * 60 * 10)))
                 .signWith(SignatureAlgorithm.HS256, SECRET_KEY).compact();
     }
